+++ tests/jest.ts
@@ -1,6 +1,0 @@
-import { testCaseFileExistence, testEnumsOverlapping } from "./shared/test-consistency";
-
-export default () => {
-  testEnumsOverlapping();
-  testCaseFileExistence();
-};