+++ tests/shared/enum.ts
@@ -1,0 +1,22 @@
+export enum TestCaseDescribe {
+  arrowFunctionCase = "arrow function case",
+  classicFunctionCase = "classic function case",
+}
+export enum TestCommonDescribe {
+  noFunctionCase = "no function case",
+}
+export enum TestResult {
+  functionError = "function error result",
+  functionSuccess = "function success result",
+}
+export enum TestCaseIt {
+  callbackEndingStrategy = "callback ending strategy",
+  synchronousStrategy = "synchronous strategy",
+  promiseStrategy = "promise async strategy",
+}
+export enum TestCommonIt {
+  getNoFunctionalPropertyStrategy = "get no functional property strategy",
+}
+export enum Prefix {
+  commonPrefix = "common prefix",
+}